Add cross-origin GM_xmlHttpRequest support.
QtWebengine doesn't support cross origin XMLHttpRequests so our thin js
GM_* wrapper doesn't match up with the greasemonkey spec in that
instance[1]. This change uses QWebChannel to proxy the requests back to
the python browser instance to be performed in a more privileged
environment. I use QNetworkRequest and QNetworkAccessManager to do the
requests to fit with all the other QT stuff we are using. Sadly,
QtWebEngine uses its own network stack and we don't have access to in a
way that would let us piggyback on its request factories and access
controls. So to get access to the cookies we have to have a copy of them
in python. Also copying user-agent from the current profile.

There is still a bunch of edge behaviours to check and I want to handle
at least HTTP Basic Auth.

Currently playing with a problem where the first load of a new domain in
an existing tab results in the global `qt` object (the websocket
transport provided by qtwebengine) not being present when the
QWebChannel.js script tries to access it. It is there on a refresh
though. I don't see anything on my side that could be doing it (there
are some notes in the docs about ensuring the webchannel is registered
before the channel is used) so I'll have to have a poke around in the
C++.

[1]: https://wiki.greasespot.net/GM_xmlhttpRequest