Don't save _autosave session on exit.
If a saveable is created without a `changed` signal passed in then it
will have `_save_on_exit` set. Which causes it to be saved in
`Quitter._shutdown()` after all the windows have shutdown. It is deleted
at the end of that function but if for any reason it wasn't (there are
already workarounds for potential crashes in that area) then an empty
session would be "recovered" on next start.

So we pass a signal into the saveable to stop that. The signal should be
redundant given we are marking it as dirty in `on_load_finished` anyway
but if the autosave time is handled between then and when the filtered
signal is emitted it is possible it could spuriously be marked dirty.

Alternately we could add `save_on_exit` flags to `add_saveable()` and
`Saveable.__init__()` to allow overriding the default behaviour.