Add tests for statusbar settings widget.
I tried to go the opposite direction than most of the tests I have
contributed previously. These ones try to be proper unit tests by
testing one thing and mocking out everything else. It makes them very
uninteresting but sometimes the mocking involves a few contortions.

I found it surprising that a window was constructed when running unit
tests, this was caused by using the `fake_statusbar` mock as parent of
the unit under test. Without using that I still had to use `qtbot`
(whatever that is) or else I would get a stacktrace on init of the base
`QLabel`.

`_current_tab()` is mocked out on most of the tests to avoid having to
depend on a window registry and a fake window and tab.

I didn't bother testing weird stuff like passing an int in where a
string is expected because the config system just doesn't allow it in
real usage.

TODO: add tests for test_feature on the tabs too.