Use an Enum to represent feature permission state.
This changes `AbstractPermissions.test_feature()` to now return an
instance of `shared.FeatureState`. The possible values of that enum
happen to have the same values as the BoolAsk ConfigType which is useful
for constructing a return type.

Thus `test_feature` will no longer return `False` for features in the
`"ask"` state, it will now return `FeatureState.ask`.

Also I forgot to take the `setting_name = "<unknown>"` fallback for
webkit out in a previous commit so do so here.