Don't break mouse back event on webengine.
AbstractTabHistory.{back,forward} raises a WebTabError which is handled
by the commands :back/:forward but not by the mouse event handler
(qutebrowser/browser/mouse.py). It instead calls `can_go_{back,forward}`
to check whether it should bother calling the history api. For some
reason 'can_go_back()` returns True on webengine when the tab is
unloaded (it should have no history in that case). It returns false for
forward and for both of them on webkit.