Add basic extension loader.
This is similar to `_walk_normal()` but since the extensions aren't in a
properly impartable package we have to actually use the finder returned
by `walk_packages()` to load the source files. That means the module is
actually imported at that point and when `_load_component()` uses
importlib it finds the module alreay loaded and everything works.

TODO: error handling

* when `os.mkdir()` fails, because we have seen that before,
* and when importing the extensions, because I assume we don't want
  third party extensions to crash the browser when there dependancies
  aren't met or something,
* and when running the module init hooks for the same reason.

TODO: support loading folders

If I remove the `if ispkg: continue` check the finder line actually
succeeds in loading the directory (if it has a `__init__.py`) but after
we yield it drops back into the `walk_packages` implemention after its
yield where it calls `__import__(name)` which fails because it
presumably tries to load the parent package which isn't real.
It may be as simple as ignoring errors for packages.