Tie AbstractWebInspector to a single page.
This fits with the current usage of inspectors and allows us to dispose
of them when the page is deleted. Except on 5.7 where an inspector can
view any page so it can be a little strange to have the inspector bound
to whatever tab you had focussed when you created it.

* requires a page argument to inspector.create()
* make the inspect method private, the only externally usable method is
  toggle()

TODO: adjust tests