Stop inspecting when inspector hidden.
On the Network tab of the web inspocter the tooltip for the Disable
Cache checkbox reads

> Disable cache (while DevTools is open)

which was previously not the case since. Since we never stopped it
inspecting or destroyed it it would keep inspecting and drawing to the
hidden widget and network calls would continue to bypass the cache.

I tested it by watching the access log of an apache server with
modpagespeed.

This does making re-opening the inspector a little slower than
previously, but not as slow as opening it for the first time. And, on
webengine 5.11+ the state is gone on re-opening it like, in network
graph and stuff in the console. But that is the same behaviour as other
browsers.

This doesn't really gain anything on 5.7 because the inspector can look
at any tab.

It would be feasible destroy the child views which would save more
memory when closed but the abstract tab would
have to be adjusted for allowing it to be None sometimes.