Use copy constructer for adding cookies to cache.
I'm not quite sure when this changed but it seems the pointer we get on
cookie added becomes invalid after returning from the signal handler. As
far as I can tell the only update I did to webengine recently is from
(debian) 5.11.2+dfsg-2 to 5.11.3+dfsg-1+b1. I'm not sure why this would
change on that update.

Luckily QNetworkCookie has a copy constructer (unlike most Qt objects
which require serialization) and the equality operator still sees copies
as equivalent.

Also tested on 5.13.0 (kde-unstable) on arch and saw the same issue. It
could be that something else in my setup changed to cause it but the
only thing I have done recently is cross-graded my graphics stack from
i386 to amd64 and that was days ago. This crash would have happened on
every cookie removal at least (so on every page reload). Strange.