Support loading package extensions.
If you have a folder in your extensions folder with a `__init__.py` file
in it it will now be loaded as an extension. So will the modules under
it.

Extensions can't currently be imported from
`qutebrowser.extensions.third_party` but if we added a `third_party`
folder in qutebrowser/extentions (even an empty one) we would be able to
import them using the "from ... import <extension>" syntax.

The issue was that when pkgutil whent to import the package it was
failing because the top level module ("extensions") wasn't real. Now it
is under a real package.