Add error handling.
Handle errors importing extensions and running their initialization and
config changed hooks by logging the error (via a stack trace) and
continuing.

I put the name/filename of the extensions in the log messages but now I
realise the filename is going to be in the stack trace. If lint
complains I will just take them out.