Use now tab-local interceptor registration.
The profile registration method is now marked obsolete (for running in
the IO thread): https://doc-snapshots.qt.io/qt5-dev/qwebengineprofile-obsolete.html#setRequestInterceptor

I could have sworn I also got a warning on the console about us using
the old API too but I don't see it anymore. Not sure if it is a one time
thing or I imagined it.

There is also a new `profile.setUrlRequestInterceptor()` which doesn't
run in the IO thread. So putting the version check in the interceptor
would be simpler but I have plans for the per-page one.