Fix config changed hook error handling.
The case where they wanted to be called for any config item wasn't
covered by the broad except.

I don't want the config change filter validation to be in the broad
except but I am going to move that out in a future PR anyway. It is
curretly a memory leak (when change filters are instantiated they are
put into a module scope list, here we instantiate them whenever a config
item is set) and we can fail early on that.