Make sure background tabs report as hidden.
To make background tabs have `document.hidden` set to `true` we must:
1. trigger page load
2. show the widget
3. hide the widget

Missing any of the steps or doing them in a different order will make
the page report as visible when the page first loads. Calling
`.reload()` on it will make it correctly report as hidden. The widget
itself correctly reports it's visibility, it is just in JS. Maybe when
the page is initialised is thinks it is visible but the widget thinks it
is hidden so when you `.hide()` the widget it doesn't bother propagating
that to the page (ie, the chromium renderer) hence why we have to
`.show()` it first.

Technically we could just do `.show(); .hide()` for every tab and let
the `.setCurrentWidget()` for foreground tabs take care of foregrounding
them but I think that would be more surprising when reading the code.

You can test it with:

    :open -b javascript:window.setTimeout(function(){document.write(''+document.hidden)},0)

It should report false without this change on 5.13 and previous and report
true with this change.