Load tabs from sessions as background tabs.
To firm up the visibility life cycle of tabs that wont be focussed we
should load the in the background.

The extra `.show()` + `.hide()` is only required prior to version 5.13.
Previous to that views need to be shown then hidden after the first page
load has been triggered. On 5.13 they you can show then hide them before
the first page load, which tabopen does.