We don't need to explicitly show foreground tabs.
This was causing an issue in tests (the stacking ones in tabs.feature)
because it seems that putting the `.show()` before
`self.widget.setCurrentWidget()` was causing the widget show action to
take a little longer (about 0.376s on my machine). Possibly because the
`.show()` is a no-op when the QTabbar has already shown it.

We need to show background tabs before we hide them to synchronize the
visiblity state with the webview but that isn't the case for foreground
ones.

It it unfortunate that we have to work around timing events like this
instead of being able to, for example, wait until the event loop had no
pending events. But running in a separate process makes that diffcult.
Running the tests in the same process would mean we would have to
**not** call app.exec() and manage flushing the event queue (like
https://stackoverflow.com/a/22951970) and probably spying on all signals
ourselves.