Scroll tab bar so current tab isn't at an edge
It is unfortunate that everything we could override to fix the "gravity"
of the visible portion of the tab bar is in QTabBarPrivate. Also that it
has no public API for scrolling ... except for the buttons.

Gets the QToolButtons that are children of the QTabWidget (actually of
the QTabBarPrivate) and uses them to scroll the tab bar. Only works
because they aren't hidden, they are just of zero size.

Gets the tab indices at the edges of the tab bar and compares them to
the current tab and if there is less than 5 tabs differnce it scrolls
across to make up the difference. I didn't bother making the padding
count configurable because I doubt this'll get merged.

The throttle is mostly for resizing and shutting down (not sure why that
is not a decorator).