Don't add to to_load for new history items.
`to_load` holds history items for unloaded tabs, which don't get
navigation events so no need to add to it here. There was also no
handling of going back in history and then opening a new page. With the
backend history implementations gets rid of the future history in that
case.