Handle editable div and span with external editor.
For both of those elements the value attribute isn't as useful as with
textareas and inputs. It can be set but it doesn't reflect the actual
text content. This falls back to innerText (via webelem's __str__()) if
value doesn't have anything useful and special cases div and span to
set textContent instead of value.