wip save page state in session file
With Qt 5.15 the underlying chromium version switched to using a new
page serialization format. Additionally the deserialization is much more
fragile and we don't have enough of an API from webengine to pull out
enough of the necessary page attributes to construct something that we
can deserialize into a new page that works. So now we attempt to dump
the whole page state along with the session. This should be backwards
compatible so if you save a session with this version of qutebrowser on
5.14 and then load it on 5.15 you should still get your session history.

I have no idea how fragile the parsing is.

TODO:
cleanup
abstractisize (make work for webkit)
test on other versions and document weirdnesses
add version number to session file and save backups on loading older ones?

ref https://github.com/qutebrowser/qutebrowser/issues/5359