mypy: switch from finder.load_module to find_spec
FileFinder.find_module(name) doesn't comply with whatever stubs mypy is
looking at. The source mentions it is deprecated this one.

Still had to add a couple of unnecessary things to make mypy happy, the
`None` argument to find_spec() is defined in the source as `path=None`
but mypy thinks it is neither optional nor usable as a keyword argument.

The (fullname) argument to load_module() appears to be unnecesary in
practice but mypyp insists.