Re-write all PyQt imports to go via qutebrowser.qt
This is the first step in making qutebrowser work with both PyQt5 and
PyQt6. Namely: not having that module name everywhere throughout the
codebase.

This was a mechanical re-write performed like so:

    git ls-files -z | xargs -0 sed -i 's/^\(\s*\)from PyQt5\./\1from qutebrowser.qt./'
    git ls-files -z | xargs -0 sed -i 's/^\(\s*\)import PyQt5\./\1import qutebrowser.qt./'
    git ls-files -z | xargs -0 sed -i 's/^\(\s*\)from PyQt5 import/\1from qutebrowser.qt import/'

This commit won't run as the imports will currently fail, see the next
commit for making them work.

I was hoping to use semgrep but, while it seems to detect import
patterns fine, using it to autofix them doesn't work so well. When
multiple objects are imported from somewhere it attempts to do the
autofix once for each thing imported. So the fixes mangle each other
because they don't re-compute what they are supposed to do.

Eg `semgrep --lang=py -e 'from PyQt5.$SUBMODULE import ($IMPORTEES)' --replacement 'from qutebrowser.qt.$SUBMODULE import ($IMPORTEES)' qutebrowser/mainwindow/tabbedbrowser.py -a`

The "importees" part of the import pattern doesn't seem to be optional
either.