Import PyQt via proxy module loader.
We want to load modules from PyQt via our own wrapper so we can
transparently support both PyQt5 and PyQt6 without having conditional
imports all through the codebase. We have a few options for this:

1. fake modules/packages and do a search and replace of imports
2. import everything we need into one module and adjust imports to use
   those module attributes
3. do either 1 or 2 AND adjust all the references to Qt types to not be
   the upstream QtThing but just like "thing"

(1) is the least effort and so of course what I have implemented here.
Apart from the package stuff is imported from no code needs changing.
For example:

    from PyQt5.QtGui import QKeyEvent, QIcon, QPixmap

changes to

    from qutebrowser.qt.QtGui import QKeyEvent, QIcon, QPixmap

For (2) we would have to change that to

    from qutebrowser.qt import QtGui

And then refer to the types like QtGui.QKeyEvent, QtGui.QIcon and QtGui.QPixmap.
Automating that re-write would mean learning some new stuff for me so I
stuck with the option that can be done in an afternoon for now.

(3) just builds on (2) to not have Qt on everything. And it may make
adapting to some webengine classes moving in Qt6 easier.

This will likely make pylint, mypy and some other stuff very confused.

To get a list of all the Qt modules we use you can use semgrep, although
I didn't quite get it to re-write stuff for me:

    semgrep --lang=py -e 'from PyQt5.$SUBMODULE import ($IMPORTEES)' -o findings.json --json qutebrowser scripts/ tests/

And then to parse the json:

    with open("findings.json") as f: data=json.load(f)
    results = data['results']
    submodules = sorted(set(r['extra']['metavars']['$SUBMODULE']['abstract_content'] for r in results))
    for m in submodules: print(f"from pyqt import {m}")
    for m in submodules: print(f'{m} = importlib.import_module("PyQt5.{m}")')

Does the qt.py file need to be moved to be a package? Probably not for
this setup, I was trying to do something else at the start...

TODO:
* see if we can get the static analyses working with this
* or try using PyParsing, lib2to3, libCST, or pyupgrade to do (2) and
  get rid of the proxy loader
* undo the mechanical changes to scripts/ (leave them referring to PyQt
  directly and change them on the Qt branch
* undo the mechanical changes to the userscripts (maybe make a
  standalone version of this for them?)

Relates to #995