wip: add initial attempt at making rename codemod support wildcards
It works! With lots of printf debugging!

Now that I've gone through and made it work I need to go back and try to
understand it and think about all the cases I haven't tested and might
have broken.

This is my minimal example:

before:
    from qutebrowser.qt.QtCore import (QUrl,
                                       QPoint)
    from qutebrowser.qt.QtGui import QIcon
    foo = QUrl("about:blank")
    bar = QPoint()
    qux = QIcon()

after:
    from qutebrowser.qt import QtCore, QtGui
    foo = QtCore.QUrl("about:blank")
    bar = QtCore.QPoint()
    qux = QtGui.QIcon()

I've already ran into one case that I didn't cover with that:

    from qutebrowser.qt.QtCore import Qt

    foo = Qt.Key_Enter