rename codemod: reset cache vars between modules
(From my LibCST fork e6b990836204b)

The rename codemod currently has a few variables cached on `self` which
are used to transfer information between different visitor methods.
For example one of the nodes being visited inside an import from
statement might cause the import to be skipped in `leave_Import` using
`self.bypass_import`.
Or in the test case below relevant seen imports are cached in
`self.scheduled_removals` so they can be given special attention in
`leave_Module`.

The lifecycles don't work out though as the transforms that codemods run
ass are only initialised once for a whole run. So stuff can leak between
modules. There is probably a better place for them, but this works for
now. I'm not sure if this is reproducable outside of this wildcard
rename support.

Here is an example:

1. Have a folder with two test files

    ==> context_test/1.py <==
    from a.b import c

    c.x.foo

    ==> context_test/2.py <==
    from a.b.g import d

    d.bar

2. Rename something from the first file

    python3 -m libcst.tool codemod rename.RenameCommand context_test/ -j 1 --old_name=a.b.c.* --new_name=a.b:c.* --no-format ; head -n-0 context_test/*

3. Observe a removed import from the first file leaks into the second

    ==> context_test/1.py <==
    from a.b import c

    c.x.foo

    ==> context_test/2.py <==
    from a.b.g import d
    from a.b import c

    d.bar