Manual fixups for ImportErrors lost in rewrite
Fixups for commit Qt import rewrite commit "Re-write PyQt references to always include the module"

Imports not at the top of the file were mostly moved up there. Where
they were inside a try/except block they were replaced with `pass`
(why!).

In a previous version of this commit it was based on a branch were we
were doing things like:

    if module and module.attribute is not None:
        do_thing()
    elif hasattr(module, 'attribute'):
        do_thing()
    else:
        cant_do_thing()

But on this branch we are doing it like:

    try:
        import module.attribute
        do_thing()
    except ImportError:
        cant_do_thing()

They may have been some stuff that came over from the earlier version of
this commit (1378c09aa5de) that hasn't been migrated over. I'll get
back to it when I know what direction we are going with the qt wrapper.
Because different approaches require different handling of this (for
example if you just do import QtCore in the wrapper and then refer to
stuff as QtCore.attribute in the code you won't get any ImportErrors)

This has not currently been well tested. We might have to move some of
the conditional checks into qt.py.

I identified these pieces of code by looking for `^+ * pass$` in the
diff of the big rewrite commit.