Refine handling of invalid keys
Introduce an InvalidKeyError to KeyInfo which can then also be turned
into an error in other places.

This was mainly needed to correctly handle config validation for key strings
containing characters not in Qt.Key.

Also, reduce loglevel for the message to debug, as those seem to mostly happen
with keys qutebrowser shouldn't be concerned about anyways (see #7047).