run pylint with both PyQt 5 and 6
We have differences in names between 5 and 6. Mostly we use the wrapper
machinery with conditional in a few special places which need to know
the details. So we could probably get away with just having Qt6
installed (since we are preferring Qt6 locations for things already).

But webkit isn't in Qt6 so we need Qt5 if we are keeping that anyway.