Re-write PyQt references to always include the module
Still some manual fixes to make, see #995 and linked PRs.

This was done like:

    date ; for mod in core dbus gui network printsupport qml sql webengine webenginecore webenginewidgets webkit webkitwidgets widgets test opengl; do
      echo "renaming $mod"
      python3 -m libcst.tool codemod rename_pyqt.RenameCommand qutebrowser/ tests --old_name=qutebrowser.qt.$mod.* --new_name=qutebrowser.qt:$mod.* --no-format
    done ; date

Where the list of modules was obtained from the current
pyqt5/pqyt6/pyside2 import wrapper machinery. That list of modules to
include can also be obtained using semgrep like so (it might miss out
the QtTest module for some reason):

    semgrep --lang=py -e 'from qutebrowser.qt.$SUBMODULE import ($IMPORTEES)' -o findings.json --json qutebrowser scripts/ tests/

and then

    with open("findings.json") as f: data=json.load(f)

    results = data['results']
    submodules = sorted(set(r['extra']['metavars']['$SUBMODULE']['abstract_content'] for r in results))
    for m in submodules: print(m