Document how to set the backend for the tests.
I only had the old way save in my bash history and this one was only
mentioned in the changelog.

Also changed the heading above the new entry to be title case, which
seems to be more consistent with the other headinfs in the file.

Also remove the one remaining mention of `QUTE_BDD_WEBENGINE` since it
does nothing anymore.