handful of mypy fixes
The comment changes in qt.py might have to change once we figure out the
best way to run mypy (eg with PyQt5 and 6 at the same time or one at a
time? If one at a time how to properly make mypy ignore stuff like
AttributeErrors?).

I had to comment all the PyQt5 stuff in qt.py to get mypy to run with
just PyQt6 without casting everything to Any. Presumably it's possibly
to nudge mypy into type narrowing or whatever but I couldn't figure out
a way to make it work.

There are still lots of mypy errors, the most common ones are:
* complaining about webkit not having type anymore (since I was running
  it with just PyQt6 installed
* <somesignal> has no attribute "connect"
* no override for method with "bytes" as an arg (they expect QByteArray)
* complaining that a Qt module doesn't have an attribute when the
  attribute access is inside a 'except AttributeError' block or a 'if
  hasattr():' block
* a few places complaining about unexpected types being passed to
  functions, mainly that is _EnumValueType in
  qutebrowser/utils/debug.py, it seems `simplewrapper` is now treated
  differently? Maybe we are getting real types now?