Add pyqt5 stubs back for mypy
It seems it does lower the number of errors!

Still lots of errors that appear to be from the PyQt5 path though (eg
complaining that webenginecore doesn't have QWebEngineDownloadRequest
because we copied that over for qt5)

I'm starting to think just making our own copy of the stub files might
actually be a better option, despite it probably being hard to make work
with everyone's local setup (eg their IDEs).