wip: attempt to speed up message matching in e2e tests
py-spy was showing a bit of time spent in pattern_match().

I tried to speed that up by compiling patterns and caching them, it
didn't help.
Then I tried doing a quick string match to maybe skip the regex match,
that didn't help either.

Going back up the stack a bit I found some more useful stuff. When
matching a line we can get multiple attributes to match on, for example:
message, level, module and category.
Stopping matching at the first attribute that doesn't match is sensible.
As is exiting early if the line is too early or has already been seen.
Lastly I made `_match_data()` only treat strings as patterns for the
`message` attribute, and only do a straight string match otherwise. I'm
not sure if that's a solid assumption, I wasn't running with the whole
test suite for this.

I think some of that reduced the around `_match_data()` showed up in
py-spy, which is great. But none of it reduced the overall test time.
Way to much time is spent waiting for the process under test (or
potentially parsing log messages or doing qtbot stuff).