fix and unskip iframe search tests
follow on from cb0b51c09463

Remove the @skip annotations from the two tests.
Fix the console.log typo.
Add "contentful" text to another html file.
Change the active tab in the test assertion.

The point of this test is to make sure that searching, and acting on a search
match, works in an iframe. It's not the point of the test to test the
tabs.background behaviour. I'm not sure if the behaviour here has changed
since the tests where disabled but as long as this part of the test is stable,
and it showed a new tab opened, it should be fine.