wip: QTBUG-112142 screen capture picker test
Based on patcheset 12 from the above QTBug applied to 6.5.1. Will pick
up again when I've built Qt6 or there is a new patchset.

Applies cleanly to Qt5 but both screens and model lists are always empty.
The helper to select the primary screen works fine.
The helper must be called or else the capture request isn't handled - this is
  a change of behaviour for existing clients. Currently the primary screen is
  always captured, if they don't connect to the signal and call the helper the
  screen capturing will stop working on upgrade.
In early testing it seemed that not accepting a request caused memory to
  accumulate in the X11 process quite quickly.