Investigations with Qt6
It works! Bit more of a limited API than I was expecting, but it could
be useful. Need to think about what kind of UX we would want on our side
anyway.

For compiling PyQt6-6.5.2/ I had to:

QtSqlResult.boundValues -> change to be not const
QtBluetooth/sipQtBluetoothquint128.cpp comment out references to `data[i]`,
  not a proper fix but I don't think I care about this class