wip: try to cache and profile completion slowness
Inspired by https://github.com/qutebrowser/qutebrowser/discussions/7862
I was trying to figure out why the completion was so damn slow, and kept
blocking video playing. I ended up reducing the rows and setting a delay
(100) for now so at least it doesn't block other stuff in the gui
thread. But we should either be making the drawing way faster or moving
it into another thread.

For testing I was adding stuff to the history DB like:

    delete from CompletionHistory;
    insert into CompletionHistory select 'about:blank?' || value, 'about:blank?' || value, value from generate_series(1,300000);
    insert into CompletionHistory select 'zzzz?' || value, 'zzzz?' || value, value from generate_series(1,3000);
    insert into CompletionHistory select 'iiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiii?' || value, 'iiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiii?' || value, value from generate_series(1,3000);

Theories:

* too many items
* too big items?
* re-drawing stuff that isn't needed?
    * time things, identify what is being draw
* re-calcing stuff that isn't needed?
    * time things
* remove highlighter
* set_pattern takes way longer when matching `i` against `iiiiiiiiiiiiiiiii..`
* `_get_textdoc()` is called from in `_draw__text()` and `sizeHint()` (might
  be able to conditionally call it in sizeHint?
* nested painter saves and restores - seems to be necessary

From what I say the highlighter the drawing the document text (via the
layout) where the slowest bits. I'm not sure we can really work through
them with out current setup. There is stuff to cache, and that does help
a little bit but not enough to prevent the bad behaviours.

Could make a benchmark out of the tests in tests/unit/completion/test_completer.py

The QT stack trace while we are painting is:

    in sipQStyledItemDelegate::paint
    in QTreeView::drawRow
    in sipQTreeView::drawRow
    in QTreeView::drawTree
    in QTreeView::paintEvent
    in sipQTreeView::paintEvent
    in QWidget::event
    in QFrame::event
    in sipQTreeView::viewportEvent
    in QCoreApplicationPrivate::sendThroughObjectEventFilters
    in QApplicationPrivate::notify_helper
    in sipQApplication::notify
    in QCoreApplication::notifyInternal2
    in QCoreApplication::sendSpontaneousEvent
    in QWidgetPrivate::sendPaintEvent
    in QWidgetPrivate::drawWidget
    in QWidgetPrivate::paintSiblingsRecursive
    in QWidgetPrivate::drawWidget
    in QWidgetPrivate::paintSiblingsRecursive
    in QWidgetPrivate::paintSiblingsRecursive
    in QWidgetPrivate::drawWidget
    in QWidgetRepaintManager::paintAndFlush
    in QWidget::event
    in sipQWidget::event
    in QApplicationPrivate::notify_helper
    in sipQApplication::notify
    in QCoreApplication::notifyInternal2
    in QCoreApplication::sendEvent
    in QCoreApplicationPrivate::sendPostedEvents