wip: remove debug logging and add throttles
The debug logging was slowing stuff down significantly even when debug
logging wasn't turned on. Because the filtering is handled by the root
logger it still looks to the `log.completion` logger like it is enabled.
(I didn't try disable the ram handler.)

`_maybe_update_geometry()` looks quite slow on a flamegraph, primarily
due to `completionwidgets.sizeHint()`. So just throttle it, we shouldn't
need to show and hide the widget too often.

250ms is an arbitrary time based on the general human reaction time
number https://humanbenchmark.com/tests/reactiontime

Disable `_draw_icon()` because it was showing up on the flamegraphs but
I don't think we even use icons in this widget?