Fall back to loading default session if autosave fails
If loading the autosave session fails you can be left with a browser window
with just your start pages in it, instead of the default session that is
normally loaded on startup (assuming you have `auto_save.session = true`). It
is then quite easy to overwrite your default session if you aren't careful.

This is particularly helpful when an early crash caused an invalid autosave
session to be saved, which we later fail to load. See https://github.com/qutebrowser/qutebrowser/issues/4672#issuecomment-1869807246

If the default session then fails to load the user will be back in the current
state of having a single browser window with their start pages open.

After this there will be a `_autosave.yml.bak` file saved to disk. Although
there is nothing to make that obvious to the user.