Clean up unusable windows from failed autosave load
I previously added code to give up on loading the `_autosave` session if
that failed and go back to the main session, if there was one. But it
turns out that failing to load a session can leave a window hanging
about which hasn't been `show()`n, so you can't focus it, and has a tab
with an invalid URL, so you can't save a session with it. So we need to
clean any windows like that up. Since we are essentially rolling back a
session load we may as well go all the way and close any windows that
where created during the session load.

Adds a context manager to listen to the app for any new windows created
while loading a session, then closes them if we are falling back to
loading the main session.

This does set `_last_window_session` unfortunately. I'm not sure how
much that feature is used.

You can reproduce this case by:

* saving a default session and quitting
* copying that session file to `_autosave.yml` and changing one of the
  `url` entries in the file to be an empty string (`''`)
* re-opening qute