WIP: try opening prompt test pages in new tabs always
I think what might be going on here in the flaky tests is that we are
opening new test pages in existing tabs and clicking before the focus
proxy has changed.

When we run click-element it sends a "fake" click event by default, that
is a click event simulated with a QEvent that looks like it was sent by
a user to the browser, as opposed to clicking via JS which has certain
limitations (eg things that require user interaction to prevent
background click jacking often don't work with it).
In `AbstractTab.send_event()` we send QEvents to `view.focusProxy()` and
it looks to me like the focus proxy is changing after we send the event.
I only checked with a single poll though, I don't know if we can get an
event for focus proxy changed to be sure.

Not sure what the right fix is for this. Re-try sending events?
Arbitrary delay? We could maybe connected to the `destroyed()` signal of
the focus proxy and ??? That's problematic because sometimes it won't
get destroyed and it's hard to know when (first load? some process
models? Different Qt versions?).