WIP: try delaying send_event a little instead
From what I can tell there is a race condition where an event sent to a view's
focus proxy won't be processed if it's immediately after opening a new URL in
an existing tab. My evidence so far:

* changing all the "open in" lines to "open in new tab" lines makes the flaky
  tests (eg in prompt.feature) work
* adding a small delay with QTimer like in this commit makes the flaky tests
  work

To recap, the tests I'm focussing on are ones sending a click event (eg
`click-element`) that send a "fake" click via Qt events. We see everything up
to the click appear to work correctly but we never see the effects of the
element on the page being clicked.

I've been testing with `python3 -m pytest
tests/end2end/features/test_prompts_bdd.py -k test_javascript_confirm`

I initially thought the issue was the RenderWidgetHostViewQt being swapped out
on the new page load and the old one being send the event. So I added a bunch
of debug logging to try to get a better view of that. None of this debug
logging is showing up when the tests fail so it seems that isn't the case.

I spent some time trying to get the end2end test process to run under GDB and
print when the got into the handleMouseEvent method so I could see if the
event was getting that far at least. I eventually managed it but the tests
don't seem to fail under gdb...

So I'm signing off for the day adding a bit of delay to send event to see if
that helps on CI like it does here.

I would prefer to handle this in more of an event based way but I don't know
what event we could use to know that the page is ready to send events to.
Even if we do just keep a delay in there, do we want to add it in all the
time? Or only soon after a page load? Or instead or putting a delay here, do
we want to add it in the test files? (Not sure how to do that in a non-error
prone way. Currently they just run :click-element, I don't want people to have
to manually remember to insert a delay before it).

gdb notes:

Changed `_executable_args()` in quteprocess to return `"gdb", "-q -x
handleMouseEvent-breakpoint -args".split() + [executable] + args` where
`handleMouseEvent-breakpoint` is:

    set breakpoint pending on
    set debuginfod enabled off
    break RenderWidgetHostViewQtDelegateClient::handleMouseEvent
      commands
        call (void) fprintf(stderr, "Entering handleMouseEvent\n")
        continue
      end
    run

The tests were consistently passing, but erroring on the invalid gdb log lines.